[NETIO] NETIO.SYS driver
This patch adds the NETIO.SYS driver to ReactOS. It is not
feature complete (meaning some functionality is unimplemented)
but does its job quite good for what it originally was written
for (which is getting WinDRBD working on ReactOS/Windows 2003 SP2).

The driver re-uses parts of the AFD.SYS driver, namely those
functions that ease communitating with the transport device
interface (TDI). Other than that, following features are implemented
and should work:

   TCP/IP networking: connect, listen, accept, read, write
   UDP/IP networking: write

So in a nutshell TCP/IP support is completed, UDP support is
partially complete and ICMP support does not exist yet.
In particular the listen/accept mechanism allows one to write
kernel side TCP servers that one connect to via the internet.