NETIO.SYS: Fixed completion routine not called.
The completion routine for the driver that called us was
never called. The reason was that nobody ever called
IoCallDriver() on the Irp of that driver. I now implemented
a dummy handler for the irp and call IoCallDriver before
doing anything else. The handler that is called just
sets the status field to STATUS_PENDING which makes sense
because the interesting things happen later after IoCallDriver
returned.